Extract into more OOP pattern
